# [PoC] sdk/log: Record is concurrent safe