--- a/sdk/log/logtest/factory_test.go
+++ b/sdk/log/logtest/factory_test.go
@@ -107,7 +107,7 @@ func TestRecordFactoryMultiple(t *testing.T) {
 	assert.Equal(t, scope, record1.InstrumentationScope())
 }
 
-func assertBody(t *testing.T, want log.Value, r sdklog.Record) {
+func assertBody(t *testing.T, want log.Value, r *sdklog.Record) {
 	t.Helper()
 	got := r.Body()
 	if !got.Equal(want) {
@@ -115,7 +115,7 @@ func assertBody(t *testing.T, want log.Value, r sdklog.Record) {
 	}
 }
 
-func assertAttributes(t *testing.T, want []log.KeyValue, r sdklog.Record) {
+func assertAttributes(t *testing.T, want []log.KeyValue, r *sdklog.Record) {
 	t.Helper()
 	var got []log.KeyValue
 	r.WalkAttributes(func(kv log.KeyValue) bool {