--- a/sdk/log/processor.go
+++ b/sdk/log/processor.go
@@ -25,10 +25,7 @@ type Processor interface {
 	// implement any retry logic. All errors returned by this function are
 	// considered unrecoverable and will be reported to a configured error
 	// Handler.
-	//
-	// Before modifying a Record, the implementation must use Record.Clone
-	// to create a copy that shares no state with the original.
-	OnEmit(ctx context.Context, record Record) error
+	OnEmit(ctx context.Context, record *Record) error
 	// Enabled returns whether the Processor will process for the given context
 	// and record.
 	//
@@ -43,10 +40,7 @@ type Processor interface {
 	// state. An implementation should default to returning true for an
 	// indeterminate state, but may return false if valid reasons in particular
 	// circumstances exist (e.g. performance, correctness).
-	//
-	// Before modifying a Record, the implementation must use Record.Clone
-	// to create a copy that shares no state with the original.
-	Enabled(ctx context.Context, record Record) bool
+	Enabled(ctx context.Context, record *Record) bool
 	// Shutdown is called when the SDK shuts down. Any cleanup or release of
 	// resources held by the exporter should be done in this call.
 	//