--- a/sdk/log/record.go
+++ b/sdk/log/record.go
@@ -42,9 +42,13 @@ func putIndex(index map[string]int) {
 }
 
 // Record is a log record emitted by the Logger.
+//
+// Record is safe for concurrent use by multiple goroutines.
 type Record struct {
 	// Do not embed the log.Record. Attributes need to be overwrite-able and
 	// deep-copying needs to be possible.
+
+	mu sync.Mutex
 
 	timestamp         time.Time
 	observedTimestamp time.Time
@@ -100,66 +104,88 @@ func (r *Record) setDropped(n int) {
 
 // Timestamp returns the time when the log record occurred.
 func (r *Record) Timestamp() time.Time {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.timestamp
 }
 
 // SetTimestamp sets the time when the log record occurred.
 func (r *Record) SetTimestamp(t time.Time) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.timestamp = t
 }
 
 // ObservedTimestamp returns the time when the log record was observed.
 func (r *Record) ObservedTimestamp() time.Time {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.observedTimestamp
 }
 
 // SetObservedTimestamp sets the time when the log record was observed.
 func (r *Record) SetObservedTimestamp(t time.Time) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.observedTimestamp = t
 }
 
 // Severity returns the severity of the log record.
 func (r *Record) Severity() log.Severity {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.severity
 }
 
 // SetSeverity sets the severity level of the log record.
 func (r *Record) SetSeverity(level log.Severity) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.severity = level
 }
 
 // SeverityText returns severity (also known as log level) text. This is the
 // original string representation of the severity as it is known at the source.
 func (r *Record) SeverityText() string {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.severityText
 }
 
 // SetSeverityText sets severity (also known as log level) text. This is the
 // original string representation of the severity as it is known at the source.
 func (r *Record) SetSeverityText(text string) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.severityText = text
 }
 
 // Body returns the body of the log record.
 func (r *Record) Body() log.Value {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.body
 }
 
 // SetBody sets the body of the log record.
 func (r *Record) SetBody(v log.Value) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.body = v
 }
 
 // WalkAttributes walks all attributes the log record holds by calling f for
 // each on each [log.KeyValue] in the [Record]. Iteration stops if f returns false.
 func (r *Record) WalkAttributes(f func(log.KeyValue) bool) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	for i := 0; i < r.nFront; i++ {
-		if !f(r.front[i]) {
+		if !f(r.front[i]) { // Warning: this can lead to deadlock
 			return
 		}
 	}
 	for _, a := range r.back {
-		if !f(a) {
+		if !f(a) { // Warning: this can lead to deadlock
 			return
 		}
 	}
@@ -167,7 +193,10 @@ func (r *Record) WalkAttributes(f func(log.KeyValue) bool) {
 
 // AddAttributes adds attributes to the log record.
 func (r *Record) AddAttributes(attrs ...log.KeyValue) {
-	n := r.AttributesLen()
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
+	n := r.nFront + len(r.back)
 	if n == 0 {
 		// Avoid the more complex duplicate map lookups below.
 		var drop int
@@ -275,6 +304,9 @@ func (r *Record) addAttrs(attrs []log.KeyValue) {
 
 // SetAttributes sets (and overrides) attributes to the log record.
 func (r *Record) SetAttributes(attrs ...log.KeyValue) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
 	var drop int
 	attrs, drop = dedup(attrs)
 	r.setDropped(drop)
@@ -326,42 +358,58 @@ func dedup(kvs []log.KeyValue) (unique []log.KeyValue, dropped int) {
 
 // AttributesLen returns the number of attributes in the log record.
 func (r *Record) AttributesLen() int {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.nFront + len(r.back)
 }
 
 // DroppedAttributes returns the number of attributes dropped due to limits
 // being reached.
 func (r *Record) DroppedAttributes() int {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.dropped
 }
 
 // TraceID returns the trace ID or empty array.
 func (r *Record) TraceID() trace.TraceID {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.traceID
 }
 
 // SetTraceID sets the trace ID.
 func (r *Record) SetTraceID(id trace.TraceID) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.traceID = id
 }
 
 // SpanID returns the span ID or empty array.
 func (r *Record) SpanID() trace.SpanID {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.spanID
 }
 
 // SetSpanID sets the span ID.
 func (r *Record) SetSpanID(id trace.SpanID) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.spanID = id
 }
 
 // TraceFlags returns the trace flags.
 func (r *Record) TraceFlags() trace.TraceFlags {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	return r.traceFlags
 }
 
 // SetTraceFlags sets the trace flags.
 func (r *Record) SetTraceFlags(flags trace.TraceFlags) {
+	r.mu.Lock()
+	defer r.mu.Unlock()
 	r.traceFlags = flags
 }
 
@@ -383,10 +431,13 @@ func (r *Record) InstrumentationScope() instrumentation.Scope {
 
 // Clone returns a copy of the record with no shared state. The original record
 // and the clone can both be modified without interfering with each other.
-func (r *Record) Clone() Record {
-	res := *r
+func (r *Record) Clone() *Record {
+	r.mu.Lock()
+	defer r.mu.Unlock()
+	res := *r //nolint // Mutex is reinitialized below.
+	res.mu = sync.Mutex{}
 	res.back = slices.Clone(r.back)
-	return res
+	return &res
 }
 
 func (r *Record) applyAttrLimits(attr log.KeyValue) log.KeyValue {