… sdk/log/exporter_test.go
… sdk/log/logger_norace_test.go
… sdk/log/simple_test.go