@MrAlias 2024-06-04
Implementation PoC looks good.

The performance impact is going to be sever, as you mention. Ideally we don't go in this direction.

Another thing not included, the `Record` cannot be copied after it is first used now. This will cause issues with user code eventually.