@pellared 2024-06-05
Another problem is that the usage of `WalkAttributes` can lead to a deadlock or even panic. E.g. if one to tries to use any of the record's method in the function passed to `WalkAttributes` it result in a panic (recursive lock).